+++ shipgen/GenLinkDef.h
@@ -15,6 +15,5 @@
 #pragma link C++ class  MuDISGenerator+;
 #pragma link C++ class  NuageGenerator+;
 #pragma link C++ class  FixedTargetGenerator+;
-#pragma link C++ class  ALPACAGenerator+;
 #pragma link C++ class  EvtCalcGenerator+;
 #endif